% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aides-package.R
\docType{package}
\name{aides-package}
\alias{aides}
\alias{aides-package}
\title{aides: Aides for Pooled Analysis of Aggregate Data}
\description{
A supportive collection of functions for pooled analysis of aggregate data. The current version supports users to test assumptions before relevant analysis of bias from study size and sequential analysis such as mentioned by Wetterslev, J., Jakobsen, J. C., & Gluud, C. (2017) \doi{10.1186/s12874-017-0315-7}.
}
\author{
\strong{Maintainer}: Enoch Kang \email{y.enoch.kang@gmail.com} (\href{https://orcid.org/0000-0002-4903-942X}{ORCID}) [translator]

}
\keyword{internal}
