% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_observed_sequential_analysis.R
\name{DoOSA}
\alias{DoOSA}
\title{Observed sequential analysis.}
\usage{
DoOSA(
  data = NULL,
  source = NULL,
  time = NULL,
  n = NULL,
  es = NULL,
  se = NULL,
  r1 = NULL,
  m1 = NULL,
  sd1 = NULL,
  n1 = NULL,
  r2 = NULL,
  m2 = NULL,
  sd2 = NULL,
  n2 = NULL,
  group = c("Group 1", "Group 2"),
  ref = 2,
  prefer = "small",
  measure = "ES",
  model = "random",
  method = "DL",
  pooling = "IV",
  trnsfrm = "logit",
  poolProp = "IV",
  alpha = 0.05,
  beta = 0.2,
  anchor = NULL,
  adjust = "D2",
  plot = FALSE,
  SAP = FALSE
)
}
\arguments{
\item{data}{DATAFRAME consists of relevant information.}

\item{source}{CHARACTER for labeling the included data sets.}

\item{time}{NUMERIC values of time sequence.}

\item{n}{INTEGER values of sample sizes.}

\item{es}{NUMERIC values of effect sizes.}

\item{se}{NUMERIC values of standard errors for the effect sizes.}

\item{r1}{INTEGER values of observed events in group 1 in the included data.}

\item{m1}{NUMERIC values of estimated means in group 1 in the included data.}

\item{sd1}{NUMERIC values of standard deviations in group 1 in the
included data.}

\item{n1}{INTEGER values of sample sizes in group 1 in the included data.}

\item{r2}{INTEGER values of observed events in group 2 in the included data.}

\item{m2}{NUMERIC values of estimated means in group 2 in the included data.}

\item{sd2}{NUMERIC values of standard deviations in group 2 in the
included data.}

\item{n2}{INTEGER values of sample sizes in group 2 in the included data.}

\item{group}{CHARACTER for labeling two groups.}

\item{ref}{NUMERIC values of 1 or 2 for indicating group 1 or 2 as reference.}

\item{prefer}{CHARACTER of "small" and "large" for indicating which direction
is beneficial effect in statistic test.}

\item{measure}{CHARACTER for indicating which statistic measure should be used.}

\item{model}{CHARACTER of "random" and "fixed" for indicating whether
to use random-effects model or fixed-effect model.}

\item{method}{CHARACTER for indicating which estimator should be used in
random-effects model. In addition to the default "DL" method,
the current version also supports "REML" and "PM" methods for
calculating heterogeneity estimator.}

\item{pooling}{CHARACTER for indicating which method has to be used for pooling
binary data. Current version consists of "IV" and "MH" for
binary data pooling.}

\item{trnsfrm}{CHARACTER for indicating which method for transforming pooled
proportion. Current version supports "none", "logit", "log",
"arcsine", and "DAT" for the transformation.}

\item{poolProp}{CHARACTER for indicating which method has to be used for pooling
proportion. Current version supports "IV" and "GLMM" for the
data pooling.}

\item{alpha}{NUMERIC value between 0 to 1 for indicating the assumed type I
error.}

\item{beta}{NUMERIC value between 0 to 1 for indicating the assumed type II
error.}

\item{anchor}{NUMERIC value for indicating the presumed meaningful effect based
on anchor-based approach.}

\item{adjust}{CHARACTER for indicating how to adjust optimal information size.
Current version consists of "none", "D2", "I2", "CHL", "CHM", and
"CHH" for the adjustment.}

\item{plot}{LOGIC value for indicating whether to illustrate alpha-spending
monitoring plot.}

\item{SAP}{LOGIC value for indicating whether to show sequential-adjusted
power.}
}
\value{
\strong{DoOSA()} returns a summary on the result of sequential analysis, and can be
stored as an object in \code{DoOSA} class. Explanations of returned information are
listed as follows:

\item{studies}{Numbers of studies included in the sequential analysis.}
\item{AIS}{Acquired information size refers to the total sample size in the
sequential analysis.}
\item{alpha}{A numeric value of type I error for the sequential analysis.}
\item{beta}{A numeric value of type II error for the sequential analysis.}
\item{OES}{A numeric value of observed effect size of meta-analysis.}
\item{variance}{A numeric value of variance of meta-analysis.}
\item{diversity}{A numeric value to show diversity in the pooled analysis.}
\item{AF}{A numeric value of adjustment factor.}
\item{OIS.org}{A numeric value for optimal information size without adjustment.}
\item{OIS.adj}{A numeric value for optimal information size with adjustment.}
\item{frctn}{A vector of fraction of each study included in the sequential
analysis.}
\item{weight}{A vector of weight of each study included in the sequential
analysis.}
\item{es.cum}{A vector of cumulative effect size in the sequential analysis.}
\item{se.cum}{A vector of standard error for the cumulative effect size in the
sequential analysis.}
\item{zval.cum}{A vector of cumulative z-value in the sequential analysis.}
\item{asb}{A data frame of alpha-spending values for each study.}
\item{aslb}{A numeric value for lower alpha-spending boundary.}
\item{asub}{A numeric value for upper alpha-spending boundary.}
}
\description{
\strong{DoOSA()} is a function for conducting observed sequential analysis.
}
\details{
\enumerate{
\item Basic information for the function \strong{DoOSA()}:
\strong{DoOSA()} supports observed sequential analysis of aggregate data synthesis
based on head-to-head comparison using either binary or continuous data in
each group. Minimum information for the function \strong{DoOSA()} encompasses a data
set of study-level data, and time sequence. Operative points of using function
\strong{DoOSA()} are listed below:
}

1.1. Parameter \code{data} should be used for assigning a data set.

1.2. Study-level data have to be assigned according to outcome type:

1.2.1. \strong{For dichotomous outcome}: Parameter \code{n1} and \code{n2} should be defined
with parameter \code{r1} and \code{r2}.

1.2.2. \strong{For continuous outcome}: parameter \code{n1} and \code{n2} should be defined
with parameter \code{m1}, \code{sd1}, \code{m2}, \code{sd2}.

1.3. Parameter \code{source} and \code{time} are required for doing observed sequential
analysis. Other parameters are auxiliary.
\enumerate{
\item Default in the function \strong{DoOSA()}
Certain defaults have been elucidated in the introductory section about the
parameters, but some of them need to be elaborated upon due to their complexity.
}

2.1. Default on the parameter \code{measure} is \code{"ES"} that automatically uses risk
ratio ("RR") for binary outcome and mean difference ("MD") for continuous
outcome respectively. Argument \code{"OR"} and \code{"SMD"} can be used for the parameter
\code{measure} when original analysis pools data based on odds ratio or standardized
mean difference.

2.2. Default on the parameter \code{method} is \code{"DL"} for applying DerSimonian-Laird
heterogeneity estimator in the original pooled analysis. Other eligible arguments
for the parameter are \code{"REML"} for restricted maximum-likelihood estimator,
\code{"PM"} for Paule-Mandel estimator, \code{"ML"} for maximum-likelihood estimator,
\code{"HS"} for Hunter-Schmidt estimator, \code{"SJ"} for Sidik-Jonkman estimator,
\code{"HE"} for Hedges estimator, and \code{"EB"} for empirical Bayes estimator.

2.3. Default on the parameter \code{pooling} is \code{"IV"} for applying inverse variance
weighting method. Other commonly-used and eligible arguments for the parameter
are \code{"MH"} for Mantel-Haenszel method and \code{"Peto"} for pooling data using Peto
method. The arguments \code{"MH"} and \code{"Peto"} are exclusively available for binary
outcomes, while the argument \code{"IV"} will be automatically applied in the case
of continuous outcomes.

2.4. Default on the parameter \code{adjust} is \code{"D2"} for adjusting optimal information
size (OIS) based on diversity (D-squared statistics). Other eligible arguments
for the parameter are \code{"None"} for the OIS without adjustment, \code{"I2"} for adjusted
OIS based on I-squared statistics, \code{"CHL"} for adjusted OIS based on low
heterogeneity by multiplying 1.33, \code{"CHM"} for adjusted OIS by multiplying 2
due to moderate heterogeneity, and \code{"CHL"} for adjusted OIS by multiplying 4
due to high heterogeneity.
}
\examples{
## Not run:
# 1. Import a dataset of study by Fleiss (1993)
library(meta)
data("Fleiss1993bin")

# 2. Perform observed sequential analysis
output <- DoOSA(Fleiss1993bin, study, year,
                r1 = d.asp, n1 = n.asp,
                r2 = d.plac, n2 = n.plac,
                measure = "RR",
                group = c("Aspirin", "Control"))

## End(Not run)

}
\references{
Jennison, C., & Turnbull, B. W. (2005). Meta-analyses and adaptive group
sequential designs in the clinical development process.
\strong{Journal of biopharmaceutical statistics}, \emph{15(4)}, 537–558.
https://doi.org/10.1081/BIP-200062273.

Revicki, D., Hays, R. D., Cella, D., & Sloan, J. (2008). Recommended methods
for determining responsiveness and minimally important differences for
patient-reported outcomes. \strong{Journal of clinical epidemiology}, \emph{61(2)},
102-109.
https://doi.org/10.1016/j.jclinepi.2007.03.012.

Wetterslev, J., Jakobsen, J. C., & Gluud, C. (2017). Trial sequential analysis
in systematic reviews with meta-analysis. \strong{BMC medical research methodology},
\emph{17(1)}, 1-18.

NCSS Statistical Software (2023). \strong{Group-sequential analysis for two proportions}.
In \emph{PASS Documentation}. Available online:
https://www.ncss.com/wp-content/themes/ncss/pdf/Procedures/NCSS/Group-Sequential_Analysis_for_Two_Proportions.pdf
}
\seealso{
\code{\link{DoSA}}, \code{\link{PlotOSA}}, \code{\link{PlotPower}}
}
\author{
Enoch Kang
}
