% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_of_study_sizes.R
\name{PlotDistrSS}
\alias{PlotDistrSS}
\title{Illustrate distribution of study sizes.}
\usage{
PlotDistrSS(n, data = NULL, study = NULL, time = NULL, method = "default")
}
\arguments{
\item{n}{NUMERIC values for sample size (n) of each study.}

\item{data}{DATA FRAME consists of three columns for study label, study
year, and sample size.}

\item{study}{CHARACTER for study labels.}

\item{time}{NUMERIC values of time sequence.}

\item{method}{CHARACTER for indicating which method should be used for testing
normality.}
}
\value{
\strong{PlotDistrSS()} returns a plot of distribution of sample sizes.
}
\description{
\strong{PlotDistrSS()} is a function for illustrating graphics of distribution of study sizes.
}
\references{
Rosenblatt, M. (1956). Remarks on Some Nonparametric Estimates of a Density
Function. \strong{The Annals of Mathematical Statistics}, \emph{27(3)}, 832–837.
doi:10.1214/aoms/1177728190.

Parzen, E. (1962). On Estimation of a Probability Density Function and Mode.
\strong{The Annals of Mathematical Statistics}, \emph{33(3)}, 1065–1076.
doi:10.1214/aoms/1177704472. JSTOR 2237880.
}
\seealso{
\code{\link{TestDisparity}}, \code{\link{PlotDisparity}}
}
\author{
Enoch Kang
}
